refactor: move create_internal_tx_for_testing to utils
