+++ crates/gateway/src/starknet_api_test_utils.rs
@@ -5,8 +5,10 @@ use starknet_api::external_transaction::{
     ExternalDeployAccountTransactionV3, ExternalInvokeTransaction, ExternalInvokeTransactionV3,
     ExternalTransaction,
 };
+use starknet_api::internal_transaction::{InternalInvokeTransaction, InternalTransaction};
 use starknet_api::transaction::{
-    Calldata, ResourceBounds, ResourceBoundsMapping, TransactionSignature,
+    Calldata, InvokeTransaction, InvokeTransactionV3, ResourceBounds, ResourceBoundsMapping,
+    TransactionSignature,
 };
 
 pub const VALID_L1_GAS_MAX_AMOUNT: u64 = 1662;
@@ -17,6 +19,31 @@ pub enum TransactionType {
     Declare,
     DeployAccount,
     Invoke,
+}
+
+pub fn create_internal_tx_for_testing() -> InternalTransaction {
+    let tx = InvokeTransactionV3 {
+        resource_bounds: ResourceBoundsMapping::try_from(vec![
+            (starknet_api::transaction::Resource::L1Gas, ResourceBounds::default()),
+            (starknet_api::transaction::Resource::L2Gas, ResourceBounds::default()),
+        ])
+        .expect("Resource bounds mapping has unexpected structure."),
+        signature: Default::default(),
+        nonce: Default::default(),
+        sender_address: Default::default(),
+        calldata: Default::default(),
+        nonce_data_availability_mode: DataAvailabilityMode::L1,
+        fee_data_availability_mode: DataAvailabilityMode::L1,
+        paymaster_data: Default::default(),
+        account_deployment_data: Default::default(),
+        tip: Default::default(),
+    };
+
+    InternalTransaction::Invoke(InternalInvokeTransaction {
+        tx: InvokeTransaction::V3(tx),
+        tx_hash: Default::default(),
+        only_query: false,
+    })
 }
 
 pub fn external_tx_for_testing(
