+++ crates/gateway/tests/end_to_end_test.rs
@@ -1,10 +1,7 @@
 use mempool_infra::network_component::CommunicationInterface;
 use starknet_api::core::{ContractAddress, Nonce};
-use starknet_api::data_availability::DataAvailabilityMode;
-use starknet_api::internal_transaction::{InternalInvokeTransaction, InternalTransaction};
-use starknet_api::transaction::{
-    InvokeTransaction, InvokeTransactionV3, ResourceBounds, ResourceBoundsMapping,
-};
+use starknet_api::internal_transaction::InternalTransaction;
+use starknet_gateway::starknet_api_test_utils::create_internal_tx_for_testing;
 use starknet_mempool_types::mempool_types::{
     Account, AccountState, GatewayNetworkComponent, GatewayToMempoolMessage,
     MempoolNetworkComponent, MempoolToGatewayMessage,
@@ -14,31 +11,6 @@ use tokio::task;
 
 pub fn create_default_account() -> Account {
     Account { address: ContractAddress::default(), state: AccountState { nonce: Nonce::default() } }
-}
-
-pub fn create_internal_tx_for_testing() -> InternalTransaction {
-    let tx = InvokeTransactionV3 {
-        resource_bounds: ResourceBoundsMapping::try_from(vec![
-            (starknet_api::transaction::Resource::L1Gas, ResourceBounds::default()),
-            (starknet_api::transaction::Resource::L2Gas, ResourceBounds::default()),
-        ])
-        .expect("Resource bounds mapping has unexpected structure."),
-        signature: Default::default(),
-        nonce: Default::default(),
-        sender_address: Default::default(),
-        calldata: Default::default(),
-        nonce_data_availability_mode: DataAvailabilityMode::L1,
-        fee_data_availability_mode: DataAvailabilityMode::L1,
-        paymaster_data: Default::default(),
-        account_deployment_data: Default::default(),
-        tip: Default::default(),
-    };
-
-    InternalTransaction::Invoke(InternalInvokeTransaction {
-        tx: InvokeTransaction::V3(tx),
-        tx_hash: Default::default(),
-        only_query: false,
-    })
 }
 
 #[tokio::test]
